news-engine/requirements.txt
asttokens==2.2.1
boto3==1.28.0
botocore==1.31.0
debugpy==1.8.0
decorator==5.1.1
distlib==0.3.7
executing==1.2.0
filelock==3.12.2
ipython==8.14.0
kaleido==0.2.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.26.4
openai==1.63.2
packaging==23.1
pandas==2.1.1
Pillow==10.0.0
plotly==5.15.0
psycopg2-binary==2.9.10
Pygments==2.16.1
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.2
pyzmq==25.1.1
requests==2.31.0
s3transfer==0.6.1
setuptools==75.8.0
six==1.16.0
ssh-import-id==5.11
stack-data==0.6.2
urllib3==1.26.12
virtualenv==20.24.3
xmltodict==0.12.0
xmltojson==2.0.2
fastapi==0.122.0
uvicorn==0.24.0
httpx==0.28.1
python-dotenv==1.0.0
pydantic==2.9.2
swagger-ui-py
pycryptodome==3.19.0
# selenium==4.9.1
selenium==4.9.1
beautifulsoup4==4.13.4
fake-useragent==2.2.0
webdriver-manager==4.0.2
schedule==1.2.2
alembic==1.17.0
undetected-chromedriver==3.5.5
google-genai==1.52.0